requirements.txt
appnope==0.1.3
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.1
Django==3.2.15
django-constance==2.9.1
django-filter==21.1
django-picklefield==3.1
djangorestframework==3.13.1
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
Pygments==2.13.0
PyMySQL==1.0.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
gunicorn==20.1.0
# Pillow==7.2.0
whitenoise==6.2.0